+++ parser/src/defs.rs
@@ -893,9 +893,12 @@ primitive!(
     ///   : ⊸(⊡⇡¯△) [1_2_3 4_5_6]
     ///   :    ⍜♭⇌  [1_2_3 4_5_6]
     ///
-    /// Subscripted [range] changes the offset of the range.
-    /// ex: ⇡₁ 5
-    ///   : ⇡₂ 5
+    /// Subscripted [range] both changes the start of the range and makes the range inclusive.
+    /// ex: ⇡₂4
+    ///   : ⇡₂3
+    ///   : ⇡₂2
+    ///   : ⇡₂1
+    /// ex: ⇡₋₃3
     /// ex: ⇡₁ 2_3
     (1, Range, MonadicArray, ("range", '⇡')),
     /// Get the first row of an array